feat(core): added author field to abstract userTextSubmission and Board.JSON
Added author to the abstract class such that ForumPosts and PostComments both have an author. The
author field is set to "anonymous" in the constructor and has to be changed with a set Method. Also
update Board.JSON to work with these new fields.

Closes #48
Development-method: pair programming
Co-authored-by:
Mattias Lund Dahlin
<mattiald@ntnu.no>