feat(ui): changed MesseptionApp to start at LoginPage
PrimaryStage now gets set to LoginPage instead of FronPage. LoginPageController has a refrence to
the FrontPageScene so this is set in messeption app by a set method

BREAKING CHANGE: It is now required to log in with a valid user in order to reach the Front Page.

Closes #61
Development-method: pair programming
Co-authored-by:
Stian Jørstad Sulebak
<stianjsu@stud.ntnu.no>