fix(core, ui): the sorting of posts now stays correct after a post gets deleted
We have improved the sortPosts method in FrontPageController. We now use it every time the frontPage
is drawn with "drawPosts()". We have created a method in UiUtils that returns a comparator based on
what we are sorting by. So some of the work has been delegated there. ForumPost now implements
comparable which definds default sorting by time.

Closes #86
Development-method: pair programming
Co-authored-by: Stian Jørstad Sulebak
<stianjsu@stud.ntnu.no>