fix(integration, rest, ui): added proper feedback on new user while running app remote
Fixed a minor bug that showed wrong feedback upon creation of an illegal user. Added a new server error 403 for whenever a user tries to create a new User with invalid name or password. This has also been implemeted so that the user gets proper feedback instead of a IOException with error code 406. It now also throws an IllegalArgumentException. No longer shows this exception type in the feedback.

Closes #114
Development-method: Individual programming